+++ spec/models/host_factory_spec.rb
@@ -66,6 +66,7 @@ describe "HostFactory" do
     let(:host) { create_host[0] }
     let(:api_key) { create_host[1] }
     context "existing host" do
+      let(:options) { {annotations: {AUTHN_ANNOTATION => 'true'}} }
       it "must be owned by the host factory" do
         create_host
         host.owner = the_user
@@ -90,58 +91,7 @@ describe "HostFactory" do
       end
     end
 
-    describe 'verify create host given AUTHN_API_KEY config' do
-      context 'when CONJUR_AUTHN_API_KEY_DEFAULT is true' do
-        before do
-          allow(Rails.application.config.conjur_config).to receive(:authn_api_key_default).and_return(true)
-        end
-
-        context 'when creating host with api-key annotation true' do
-          let(:options) { {annotations: { AUTHN_ANNOTATION => true}} }
-          it { expect { host_builder.create_host }.to_not raise_error }
-        end
-
-        context 'when creating host with api-key annotation false' do
-          let(:options) { {annotations: {AUTHN_ANNOTATION => false}} }
-          it { expect { host_builder.create_host }.to_not raise_error }
-        end
-
-        context 'when creating host without api-key annotation' do
-          it { expect { host_builder.create_host }.to_not raise_error }
-        end
-      end
-
-      context 'when CONJUR_AUTHN_API_KEY_DEFAULT is false' do
-        before do
-          allow(Rails.application.config.conjur_config).to receive(:authn_api_key_default).and_return(false)
-        end
-
-        context 'when creating host with api-key annotation true' do
-          let(:options) { {annotations: {AUTHN_ANNOTATION => true}} }
-          it { expect { host_builder.create_host }.to_not raise_error }
-        end
-
-        context 'when creating host with api-key annotation false' do
-          let(:options) { {annotations: {AUTHN_ANNOTATION => false}} }
-          it { expect { host_builder.create_host }.to raise_error }
-        end
-
-        context 'when creating host with api-key annotation False capital' do
-          let(:options) { {annotations: {AUTHN_ANNOTATION => "FALSE"}} }
-          it { expect { host_builder.create_host }.to raise_error }
-        end
-
-        context 'when creating host without api-key annotation' do
-          it { expect { host_builder.create_host }.to raise_error }
-        end
-      end
-
-      context "Without validation API key is created as expected" do
-        before do
-          allow(Rails.application.config.conjur_config).to receive(:authn_api_key_default).and_return(false)
-          allow_any_instance_of(Loader::Types::Host).to receive(:future_api_key_auth_will_fail?).and_return(false)
-        end
-
+    describe 'API key is created per the annotation' do
         context 'when creating host with api-key annotation true' do
           let(:options) { {annotations: {AUTHN_ANNOTATION => 'true'}} }
           it { expect(host_builder.create_host[1]).not_to be_nil } # create_host returns [host, api_key]
@@ -165,7 +115,6 @@ describe "HostFactory" do
         context 'when creating host without api-key annotation' do
           it { expect(host_builder.create_host[1]).to be_nil } # create_host returns [host, api_key]
         end
-      end
      end
   end
 end